NET-176: move requests func to different file
